@@ -1,16 +1,16 @@
-import React from 'react'
-import ReactDOM from 'react-dom'
-import {Provider} from 'react-redux'
-import {Router} from 'react-router-dom'
-import history from './history'
-import store from './store'
-import App from './App'
+import React from 'react';
+import ReactDOM from 'react-dom/client';
+import { Provider } from 'react-redux';
+import { BrowserRouter } from 'react-router-dom';
+import store from './store';
+import App from './App';
 
-ReactDOM.render(
-  <Provider store={store}>
-    <Router history={history}>
+const root = ReactDOM.createRoot(document.getElementById('app'));
+
+root.render(
+  <BrowserRouter>
+    <Provider store={store}>
       <App />
-    </Router>
-  </Provider>,
-  document.getElementById('app')
-)
+    </Provider>
+  </BrowserRouter>
+);
